Add custom validation
