+++ API/app/model.py
@@ -1,5 +1,6 @@
 import os
-from simplevalidator import Validator
+import re
+from passlib.hash import bcrypt
 
 
 class MealType:
@@ -8,18 +9,18 @@ class MealType:
     BREAKFAST = 3
 
 class UserType:
-    USER = 1 
-    ADMIN = 2
+    CATERER = 1 
+    CUSTOMER = 2
 
 class BAM:
     def __init__(self):
         self.clear()
-
         # create default caterer
         self.post_user({
             'username': os.getenv('DEFAULT_ADMIN_USERNAME'),
             'email': os.getenv('DEFAULT_ADMIN_EMAIL'),
-            'password': os.getenv('DEFAULT_ADMIN_PASSWORD')
+            'password': os.getenv('DEFAULT_ADMIN_PASSWORD'),
+            'role': UserType.CATERER
         })
 
     @property
@@ -29,7 +30,16 @@ class BAM:
     def post_user(self, user):
         self._users_index += 1
         user['id'] = self._users_index
+        user['role'] = UserType.CUSTOMER
+        user['password'] = bcrypt.encrypt(user['password'])
+        self._users_internal[self._users_index] = {
+            'id': self._users_index,
+            'email': user['email'],
+            'password': user['password']
+        }
+        del user['password']
         self._users[self._users_index] = user
+
         return user
 
     def put_user(self, user, id):
@@ -38,8 +48,8 @@ class BAM:
     def get_user(self, id):
         return self._users.get(id, None)
 
-    def get_user_by_email(self, email):
-        for u in self._users.values():
+    def get_internal_user_by_email(self, email):
+        for u in self._users_internal.values():
             if u['email'] == email:
                 return u
 
@@ -50,18 +60,18 @@ class BAM:
         del self._users[id]
 
     def validate_user_fails(self, fields):
-        v = Validator()
-        v.make(
-            fields=fields,
-            rules={
-                'username': 'required',
-                'email': 'required|email',
-                'password': 'required|min:8',
-            },
-        )
+        if not fields['username']:
+            return True, ['Username is required']
+
+        if not fields['email']:
+            return True, ['Email is required']
+
+        if not re.match(r"^[A-Za-z0-9\.\+_-]+@[A-Za-z0-9\._-]+\.[a-zA-Z]*$", 
+                        fields['email']):
+            return True, ['Please provide a valid email']
 
         for u in self._users.values():
-            if u['email'] == user['email']:
+            if u['email'] == fields['email']:
                 return True, ['Email must be unique']
 
         return False, []
@@ -90,20 +100,19 @@ class BAM:
         del self._meals[id]
 
     def validate_meal_fails(self, fields):
-        v = Validator()
-        v.make(
-            fields=fields,
-            rules={
-                'name': 'required',
-                'cost': 'required|posinteger',
-            },
-        )
-        if v.fails():
-            return True, v.errors()
+        if not fields['name']:
+            return True, ['Name is required']
+
+        if not fields['cost']:
+            return True, ['Cost is required']
+
+        if not str(fields['cost']).isdigit():
+            return True, ['Cost must be numeric']
 
         for m in self._meals.values():
             if m['name'] == fields['name']:
                 return True, ['Meal name must be unique']
+
         return False, []
 
 
@@ -160,8 +169,8 @@ class BAM:
         self._orders[self._orders_index] = order
         return order
 
-    def put_order(self, user, id):
-        self._orders[id] = user
+    def put_order(self, order, id):
+        self._orders[id] = order
 
     def get_order(self, id):
         return self._orders.get(id, None)
@@ -234,7 +243,9 @@ class BAM:
         return False, []
 
     def clear(self):
+        self.logged_in = []
         self._users = {}
+        self._users_internal = {}
         self._users_index = 0
         self._meals = {}
         self._meals_index = 0
